+++ dom/src/main/java/com/juliewoolie/delphidom/ExtendedView.java
@@ -4,6 +4,7 @@ import com.juliewoolie.delphi.DocumentView;
 import com.juliewoolie.dom.RenderBounds;
 import org.bukkit.entity.Entity;
 import org.joml.Vector2f;
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 public interface ExtendedView extends DocumentView {
@@ -16,7 +17,7 @@ public interface ExtendedView extends DocumentView {
 
   Vector2f getCursorScreen();
 
-  Vector3f getCursorWorld();
+  Vector3d getCursorWorld();
 
   void handleEntityVisibility(Entity entity);
